fix another three MSVC 8.0 warnings
svn path=/trunk/; revision=21089

+++ gtk/voip_calls.c
@@ -2920,9 +2920,9 @@ static int sccp_calls_packet(void *ptr _U_, packet_info *pinfo, epan_dissect_t *
 		COPY_ADDRESS(&(strinfo->initial_speaker), &(pinfo->src));
 		
 		strinfo->protocol =   SP2VP(assoc->proto);
-		strinfo->start_sec=pinfo->fd->rel_ts.secs;
+		strinfo->start_sec=(gint32) pinfo->fd->rel_ts.secs;
 		strinfo->start_usec=pinfo->fd->rel_ts.nsecs;
-		strinfo->stop_sec=pinfo->fd->rel_ts.secs;
+		strinfo->stop_sec=(gint32) pinfo->fd->rel_ts.secs;
 		strinfo->stop_usec=pinfo->fd->rel_ts.nsecs;
 		
 		strinfo->selected = FALSE;
@@ -2942,7 +2942,7 @@ static int sccp_calls_packet(void *ptr _U_, packet_info *pinfo, epan_dissect_t *
 		}
 		
 		strinfo->protocol =  SP2VP(assoc->proto);
-		strinfo->stop_sec=pinfo->fd->rel_ts.secs;
+		strinfo->stop_sec=(gint32) pinfo->fd->rel_ts.secs;
 		strinfo->stop_usec=pinfo->fd->rel_ts.nsecs;
 		strinfo->last_frame_num=pinfo->fd->num;
 		++(strinfo->npackets);
